+++ lib/android/widgets/notes_loading.dart
@@ -2,8 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:notes/android/data/data.dart';
 
 class NotesLoadingAndroid extends StatefulWidget {
+  final double strokeWidth;
   const NotesLoadingAndroid({
     Key? key,
+    required this.strokeWidth,
   }) : super(key: key);
 
   @override
@@ -21,6 +23,7 @@ class _NotesLoadingAndroidState extends State<NotesLoadingAndroid> {
   @override
   Widget build(BuildContext context) {
     return CircularProgressIndicator.adaptive(
+      strokeWidth: widget.strokeWidth,
       backgroundColor: c.onBackground.withOpacity(0.2),
       valueColor: AlwaysStoppedAnimation(c.onBackground),
     );